+++ src/types/animation.rs
@@ -1,22 +1,20 @@
+use derive_more::Deref;
 use mime::Mime;
 use serde::{Deserialize, Serialize};
 
-use crate::types::PhotoSize;
+use crate::types::{FileMeta, PhotoSize};
 
 /// This object represents an animation file (GIF or H.264/MPEG-4 AVC video
 /// without sound).
 ///
 /// [The official docs](https://core.telegram.org/bots/api#animation).
 #[serde_with_macros::skip_serializing_none]
-#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize, Deref)]
 pub struct Animation {
-    /// An identifier for this file.
-    pub file_id: String,
-
-    /// Unique identifier for this file, which is supposed to be the same over
-    /// time and for different bots. Can't be used to download or reuse the
-    /// file.
-    pub file_unique_id: String,
+    /// Metadata of the animation file.
+    #[deref]
+    #[serde(flatten)]
+    pub file: FileMeta,
 
     /// A video width as defined by a sender.
     pub width: u32,
@@ -36,14 +34,12 @@ pub struct Animation {
     /// A MIME type of the file as defined by a sender.
     #[serde(with = "crate::types::non_telegram_types::mime::opt_deser")]
     pub mime_type: Option<Mime>,
-
-    /// File size in bytes.
-    #[serde(default = "crate::types::file::file_size_fallback")]
-    pub file_size: u32,
 }
 
 #[cfg(test)]
 mod tests {
+    use crate::types::FileMeta;
+
     use super::*;
 
     #[test]
@@ -65,21 +61,25 @@ mod tests {
         "mime_type":"video/gif",
         "file_size":6500}"#;
         let expected = Animation {
-            file_id: "id".to_string(),
-            file_unique_id: "".to_string(),
+            file: FileMeta {
+                file_id: "id".to_string(),
+                file_unique_id: "".to_string(),
+                file_size: 6500,
+            },
             width: 320,
             height: 320,
             duration: 59,
             thumb: Some(PhotoSize {
-                file_id: "id".to_string(),
-                file_unique_id: "".to_string(),
+                file: FileMeta {
+                    file_id: "id".to_owned(),
+                    file_unique_id: "".to_owned(),
+                    file_size: 3452,
+                },
                 width: 320,
                 height: 320,
-                file_size: 3452,
             }),
             file_name: Some("some".to_string()),
             mime_type: Some("video/gif".parse().unwrap()),
-            file_size: 6500,
         };
         let actual = serde_json::from_str::<Animation>(json).unwrap();
         assert_eq!(actual, expected)