+++ src/types/voice.rs
@@ -1,19 +1,19 @@
+use derive_more::Deref;
 use mime::Mime;
 use serde::{Deserialize, Serialize};
+
+use crate::types::FileMeta;
 
 /// This object represents a voice note.
 ///
 /// [The official docs](https://core.telegram.org/bots/api#voice).
 #[serde_with_macros::skip_serializing_none]
-#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize, Deref)]
 pub struct Voice {
-    /// Identifier for this file.
-    pub file_id: String,
-
-    /// Unique identifier for this file, which is supposed to be the same over
-    /// time and for different bots. Can't be used to download or reuse the
-    /// file.
-    pub file_unique_id: String,
+    /// Metadata of the voice file.
+    #[deref]
+    #[serde(flatten)]
+    pub file: FileMeta,
 
     /// Duration of the audio in seconds as defined by sender.
     pub duration: u32,
@@ -21,8 +21,4 @@ pub struct Voice {
     /// MIME type of the file as defined by sender.
     #[serde(with = "crate::types::non_telegram_types::mime::opt_deser")]
     pub mime_type: Option<Mime>,
-
-    /// File size in bytes.
-    #[serde(default = "crate::types::file::file_size_fallback")]
-    pub file_size: u32,
 }